Create python-serial-interface
The core idea of this python sketch is to use the serial module to interface the computer and the driver. This is mainly for logging current and frequency measurements. But the driver can also be configured to manually control the frequency through a low-pass filtered PWM channel connected via a switch to the VCO input of the CD4046. 

NOTE: I'm completely inexperienced with python so please don't expect this to work on your machine without small to severe adjustments (for example the serial port paths will be different if you're not using windows)